requirements.txt
fastapi
pyjwt
pwdlib
pwdlib[argon2]
pydantic
typing
uvicorn
httpx
pytest
pytest-cov
python-dotenv